error handler for delete actions

+++ app/Controllers/Http/EmployeeVacationsController.ts
@@ -50,7 +50,11 @@ export default class EmployeeVacationsController {
   }
   public async destroy({ request, response }: HttpContextContract) {
     let model = await Vacation.find(request.param('id'))
-    await model?.delete()
-    response.send({ status: 'ok', id: request.param('id') })
+    if (model) {
+      await model?.delete()
+      response.send({ status: 'ok', id: request.param('id') })
+    } else {
+      response.notFound({ status: 'Not found', id: request.param('id') })
+    }
   }
 }
